LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/popup-same-origin-with-same-site.https-expected.txt


FAIL same-origin document opening popup to https://127.0.0.1:9443 with COOP: "" assert_equals: name expected "" but got "same-origin_to_SAME_SITE_"
FAIL same-origin document opening popup to https://127.0.0.1:9443 with COOP: "unsafe-none" assert_equals: name expected "" but got "same-origin_to_SAME_SITE_unsafe-none"
FAIL same-origin document opening popup to https://127.0.0.1:9443 with COOP: "jibberish" assert_equals: name expected "" but got "same-origin_to_SAME_SITE_jibberish"
FAIL same-origin document opening popup to https://127.0.0.1:9443 with COOP: "same-site" assert_equals: name expected "" but got "same-origin_to_SAME_SITE_same-site"
FAIL same-origin document opening popup to https://127.0.0.1:9443 with COOP: "same-site unsafe-allow-outgoing" assert_equals: name expected "" but got "same-origin_to_SAME_SITE_same-site-unsafe-allow-outgoing"
FAIL same-origin document opening popup to https://127.0.0.1:9443 with COOP: "same-origin unsafe-allow-outgoing" assert_equals: name expected "" but got "same-origin_to_SAME_SITE_same-origin-unsafe-allow-outgoing"
FAIL same-origin document opening popup to https://127.0.0.1:9443 with COOP: "same-origin" assert_equals: name expected "" but got "same-origin_to_SAME_SITE_same-origin"
FAIL same-origin document opening popup to https://127.0.0.1:9443 with COOP: "same-origin-allow-popups" assert_equals: name expected "" but got "same-origin_to_SAME_SITE_same-origin-allow-popups"
